editoast: move BoundingBox and Zone to editoast_schemas
