+++ editoast/Cargo.toml
@@ -12,6 +12,7 @@ chrono = { version = "0.4.38", features = ["serde"] }
 derivative = "2.2.0"
 editoast_common = { path = "./editoast_common" }
 enum-map = "2.7.3"
+geojson = "*"
 geos = { version = "8.3.1", features = ["json"] }
 mvt = "0.9.0"
 postgis_diesel = { version = "2.3.1", features = ["serde"] }
