+++ editoast/editoast_schemas/Cargo.toml
@@ -8,8 +8,10 @@ license = "LGPL-3.0"
 derivative.workspace = true
 editoast_common.workspace = true
 enum-map.workspace = true
+geojson.workspace = true
 serde.workspace = true
 serde_derive.workspace = true
 serde_json.workspace = true
 strum.workspace = true
+thiserror.workspace = true
 utoipa.workspace = true
