+++ editoast/editoast_schemas/src/errors.rs
@@ -1,0 +1,7 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum GeometryError {
+    #[error("expected geometry {expected} but got {actual}")]
+    UnexpectedGeometry { expected: String, actual: String },
+}
