+++ editoast/src/views/infra/lines.rs
@@ -9,12 +9,12 @@ use thiserror::Error;
 use crate::error::Result;
 use crate::infra_cache::InfraCache;
 use crate::infra_cache::ObjectCache;
-use crate::map::Zone;
 use crate::modelsv2::prelude::*;
 use crate::modelsv2::Infra;
 use crate::views::infra::InfraApiError;
 use crate::views::infra::InfraIdParam;
 use crate::DbPool;
+use editoast_schemas::primitives::Zone;
 
 crate::routes! {
     "/lines/{line_code}/bbox" => {
